+++ FE/src/pages/MyPage.tsx
@@ -7,6 +7,7 @@ import MyPageDiary from "@components/mypage/details/diary/Diary";
 import MyPageFrame from "@components/mypage/MyPageFrame";
 import MyPageInfo from "@components/mypage/details/myInfo/MyInfo";
 import MyFriend from "@components/mypage/details/friend/MyFriend";
+import MyPassword from "@components/mypage/details/myInfo/MyPassword";
 import { settingApi } from "@api/Api";
 import { useAppDispatch, useAppSelector } from "@hooks/hook";
 import { selectUserId } from "@store/authSlice";
@@ -70,6 +71,7 @@ const MyPage = () => {
           {
             {
               "내 정보 확인": <MyPageInfo />,
+              "비밀번호 변경": <MyPassword />,
               "다이어리 설정": <MyPageDiary />,
               "카테고리 설정": <MyPageFrame title="카테고리" />,
               "디데이 설정": <MyPageFrame title="디데이" />,
